Mejoras en el main  (#6)
* Cambios  en usuario

* -Cambiar el Home
-Modificar producto, rol y usuario

* Editar home, producto , rol y usuario

* Modificar detalles

* DetalleFactura y factura editado

* Factuas detalle y dtos

* cambios

* ambios

* mejora de facturas

* Profiles completos

* Detalle factura

* cambios en facturas// y tipos

* detalle facturas tipos

* arreglos en la migracion

* Deatelle factura

* Fondo y main cambios

* Mejoras en Layout

+++ ArtInk/ArtInk.Infraestructure/Models/DetalleFactura.cs
@@ -5,9 +5,9 @@ namespace ArtInk.Infraestructure.Models;
 
 public partial class DetalleFactura
 {
-    public short Id { get; set; }
+    public long Id { get; set; }
 
-    public short IdFactura { get; set; }
+    public long IdFactura { get; set; }
 
     public byte IdServicio { get; set; }
